# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Speed dating chatroom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history:
    get:
      operationId: getHistory
      tags: [rooms]
      summary: Get current user's game history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:

            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: matchStatus
          in: query
          schema:
            type: string
            enum: [matched, unmatched, unknown]
        - name: role
          in: query
          schema:
            type: string
            enum: [chooser, suitor]
        - name: isWinner
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest]
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated game history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /history/{gameId}:
    get:
      operationId: getHistoryGame
      tags: [rooms]
      summary: Get archived game detail for current user
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archived game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryGameDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /history/{gameId}/replay:
    get:
      operationId: getHistoryReplay
      tags: [rooms]
      summary: Get round-by-round replay timeline for an archived game
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay timeline payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryReplay"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /players/{id}/stats:
    get:
      operationId: getPlayerStats
      tags: [users]
      summary: Get derived player statistics from archived games
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /players/{id}/achievements:
    get:
      operationId: getPlayerAchievements
      tags: [users]
      summary: Get derived player achievements
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player achievements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AchievementListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [users]
      summary: Get analytics summary for dashboard reads
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 365
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /analytics/events:
    get:
      operationId: getAnalyticsEvents
      tags: [users]
      summary: Get analytics event stream for dashboard reads
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Analytics event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages, hasMore]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: ["string", "null"]
    HistoryItem:
      type: object
      required: [gameId, completionTimestamp, role, isWinner, matchStatus, roundsSurvived, winnerId, winnerName, participantCount]
      properties:
        gameId:
          type: string
        completionTimestamp:
          type: string
        role:
          type: string
        isWinner:
          type: boolean
        matchStatus:
          type: string
          enum: [matched, unmatched, unknown]
        roundsSurvived:
          type: integer
        winnerId:
          type: ["string", "null"]
        winnerName:
          type: ["string", "null"]
        participantCount:
          type: integer
    HistoryListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    ArchivedParticipant:
      type: object
      required: [id, name, role, suitorSlot, isBot, isPremium]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        suitorSlot:
          type: ["integer", "null"]
        isBot:
          type: boolean
        isPremium:
          type: boolean
...
    HistoryGameDetail:
      type: object
      required: [gameId, completionTimestamp, winnerId, winnerName, matchStatus, eliminationOrder, participants, questionsAsked, roundDurations, analyticsMetadata, questionTelemetrySummary, playerPerspective]
      properties:
        gameId:
          type: string
        completionTimestamp:
          type: string
        winnerId:
          type: ["string", "null"]
        winnerName:
          type: ["string", "null"]
        matchStatus:
          type: string
          enum: [matched, unmatched, unknown]
        eliminationOrder:
          type: array
          items:
            type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ArchivedParticipant"
        questionsAsked:
          type: array
          items:
            $ref: "#/components/schemas/ArchivedQuestion"
        roundDurations:
          type: array
          items:
            type: integer
        analyticsMetadata:
          type: object
          additionalProperties: true
        questionTelemetrySummary:
          $ref: "#/components/schemas/QuestionTelemetrySummary"
        playerPerspective:
          type: object
          required: [role, isWinner, roundsSurvived]
          properties:
            role:
              type: string
            isWinner:
              type: boolean
            roundsSurvived:
              type: integer
...
    HistoryReplay:
      type: object
      required: [gameId, completionTimestamp, winnerId, winnerName, matchStatus, rounds, eliminationOrder, participants, playerPerspective]
      properties:
        gameId:
          type: string
        completionTimestamp:
          type: string
        winnerId:
          type: ["string", "null"]
        winnerName:
          type: ["string", "null"]
        matchStatus:
          type: string
          enum: [matched, unmatched, unknown]
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/HistoryReplayRound"
        eliminationOrder:
          type: array
          items:
            type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ArchivedParticipant"
        playerPerspective:
          type: object
          required: [role, isWinner, roundsSurvived]
          properties:
            role:
              type: string
            isWinner:
              type: boolean
            roundsSurvived:
              type: integer
...
    PlayerStats:
      type: object
      required: [gamesPlayed, wins, matches, winPercentage, matchPercentage, averageFinish, averageResponseTimeSeconds, streaks]
      properties:
        gamesPlayed:
          type: integer
        wins:
          type: integer
        matches:
          type: integer
        winPercentage:
          type: number
        matchPercentage:
          type: number
        averageFinish:
          type: number
        averageResponseTimeSeconds:
          type: ["number", "null"]
        streaks:
          $ref: "#/components/schemas/PlayerStreaks"
...
    AchievementListResponse:
      type: object
      required: [achievements]
      properties:
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
...
    AnalyticsSummary:
      type: object
      required: [days, totalEvents, uniqueRooms, uniqueUsers, eventsByType, topEventTypes]
      properties:
        days:
          type: integer
        totalEvents:
          type: integer
        uniqueRooms:
          type: integer
        uniqueUsers:
          type: integer
        eventsByType:
          type: object
          additionalProperties:
            type: integer
        topEventTypes:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsTopEventType"
...
    AnalyticsEventsResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsEventItem"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [unauthorized, forbidden, not_found, invalid_query, internal_error]
        details:
          type: object
          additionalProperties: true